github-workflow: Remove quick-install test
The current quick-install tests deployment of the quickinstall yaml
without building new images and then runs the connectivity check. THis
is problematic as it means that no change will pass the test where a
helm template change depends on a code change.

The connectivity test is already being run in the "conformance-test"
test where images are being built properly.

There is a small diff in the options bein used in the "conformance-test"
and the default options in the quickinstall YAML but nothing that
wouldn't be covered by the overall CI anyway.

Signed-off-by: Thomas Graf <thomas@cilium.io>